♻️ Refactor _compat, remove Pydantic v1, first pass
