+++ fastapi/_compat/__init__.py
@@ -6,7 +6,6 @@ from .main import UndefinedType as UndefinedType
 from .main import Url as Url
 from .main import Validator as Validator
 from .main import _get_model_config as _get_model_config
-from .main import _is_error_wrapper as _is_error_wrapper
 from .main import _is_model_class as _is_model_class
 from .main import _is_model_field as _is_model_field
 from .main import _is_undefined as _is_undefined
@@ -29,11 +28,7 @@ from .main import serialize_sequence_value as serialize_sequence_value
 from .main import (
     with_info_plain_validator_function as with_info_plain_validator_function,
 )
-from .may_v1 import CoreSchema as CoreSchema
-from .may_v1 import GetJsonSchemaHandler as GetJsonSchemaHandler
-from .may_v1 import JsonSchemaValue as JsonSchemaValue
 from .may_v1 import _normalize_errors as _normalize_errors
-from .model_field import ModelField as ModelField
 from .shared import PYDANTIC_V2 as PYDANTIC_V2
 from .shared import PYDANTIC_VERSION_MINOR_TUPLE as PYDANTIC_VERSION_MINOR_TUPLE
 from .shared import annotation_is_pydantic_v1 as annotation_is_pydantic_v1
@@ -47,3 +42,4 @@ from .shared import (
 from .shared import lenient_issubclass as lenient_issubclass
 from .shared import sequence_types as sequence_types
 from .shared import value_is_sequence as value_is_sequence
+from .v2 import ModelField as ModelField
