+++ fastapi/_compat/main.py
@@ -12,9 +12,9 @@ from pydantic import BaseModel
 from typing_extensions import Literal
 
 from . import v2
-from .model_field import ModelField
 from .v2 import BaseConfig as BaseConfig
 from .v2 import FieldInfo as FieldInfo
+from .v2 import ModelField
 from .v2 import PydanticSchemaGenerationError as PydanticSchemaGenerationError
 from .v2 import RequiredParam as RequiredParam
 from .v2 import Undefined as Undefined
@@ -23,6 +23,7 @@ from .v2 import Url as Url
 from .v2 import Validator as Validator
 from .v2 import evaluate_forwardref as evaluate_forwardref
 from .v2 import get_missing_field_error as get_missing_field_error
+from .v2 import get_model_fields as get_model_fields
 from .v2 import (
     with_info_plain_validator_function as with_info_plain_validator_function,
 )
@@ -30,109 +31,50 @@ from .v2 import (
 
 @lru_cache
 def get_cached_model_fields(model: type[BaseModel]) -> list[ModelField]:
-    if lenient_issubclass(model, may_v1.BaseModel):
-        from fastapi._compat import v1
-
-        return v1.get_model_fields(model)  # type: ignore[arg-type,return-value]
-    else:
-        from . import v2
-
-        return v2.get_model_fields(model)  # type: ignore[return-value]
+    return get_model_fields(model)  # type: ignore[return-value]
 
 
 def _is_undefined(value: object) -> bool:
-    if isinstance(value, may_v1.UndefinedType):
-        return True
-
     return isinstance(value, v2.UndefinedType)
 
 
 def _get_model_config(model: BaseModel) -> Any:
-    if isinstance(model, may_v1.BaseModel):
-        from fastapi._compat import v1
-
-        return v1._get_model_config(model)
-
     return v2._get_model_config(model)
 
 
 def _model_dump(
     model: BaseModel, mode: Literal["json", "python"] = "json", **kwargs: Any
 ) -> Any:
-    if isinstance(model, may_v1.BaseModel):
-        from fastapi._compat import v1
-
-        return v1._model_dump(model, mode=mode, **kwargs)
-
     return v2._model_dump(model, mode=mode, **kwargs)
 
 
-def _is_error_wrapper(exc: Exception) -> bool:
-    if isinstance(exc, may_v1.ErrorWrapper):
-        return True
-
-    return isinstance(exc, v2.ErrorWrapper)
-
-
 def copy_field_info(*, field_info: FieldInfo, annotation: Any) -> FieldInfo:
-    if isinstance(field_info, may_v1.FieldInfo):
-        from fastapi._compat import v1
-
-        return v1.copy_field_info(field_info=field_info, annotation=annotation)
-
     return v2.copy_field_info(field_info=field_info, annotation=annotation)
 
 
 def create_body_model(
     *, fields: Sequence[ModelField], model_name: str
 ) -> type[BaseModel]:
-    if fields and isinstance(fields[0], may_v1.ModelField):
-        from fastapi._compat import v1
-
-        return v1.create_body_model(fields=fields, model_name=model_name)
-
     return v2.create_body_model(fields=fields, model_name=model_name)  # type: ignore[arg-type]
 
 
 def get_annotation_from_field_info(
     annotation: Any, field_info: FieldInfo, field_name: str
 ) -> Any:
-    if isinstance(field_info, may_v1.FieldInfo):
-        from fastapi._compat import v1
-
-        return v1.get_annotation_from_field_info(
-            annotation=annotation, field_info=field_info, field_name=field_name
-        )
-
     return v2.get_annotation_from_field_info(
         annotation=annotation, field_info=field_info, field_name=field_name
     )
 
 
 def is_bytes_field(field: ModelField) -> bool:
-    if isinstance(field, may_v1.ModelField):
-        from fastapi._compat import v1
-
-        return v1.is_bytes_field(field)
-
     return v2.is_bytes_field(field)  # type: ignore[arg-type]
 
 
 def is_bytes_sequence_field(field: ModelField) -> bool:
-    if isinstance(field, may_v1.ModelField):
-        from fastapi._compat import v1
-
-        return v1.is_bytes_sequence_field(field)
-
     return v2.is_bytes_sequence_field(field)  # type: ignore[arg-type]
 
 
 def is_scalar_field(field: ModelField) -> bool:
-    if isinstance(field, may_v1.ModelField):
-        from fastapi._compat import v1
-
-        return v1.is_scalar_field(field)
-
     return v2.is_scalar_field(field)  # type: ignore[arg-type]
 
 
@@ -141,37 +83,15 @@ def is_scalar_sequence_field(field: ModelField) -> bool:
 
 
 def is_sequence_field(field: ModelField) -> bool:
-    if isinstance(field, may_v1.ModelField):
-        from fastapi._compat import v1
-
-        return v1.is_sequence_field(field)
-
     return v2.is_sequence_field(field)  # type: ignore[arg-type]
 
 
 def serialize_sequence_value(*, field: ModelField, value: Any) -> Sequence[Any]:
-    if isinstance(field, may_v1.ModelField):
-        from fastapi._compat import v1
-
-        return v1.serialize_sequence_value(field=field, value=value)
-
     return v2.serialize_sequence_value(field=field, value=value)  # type: ignore[arg-type]
 
 
 def get_compat_model_name_map(fields: list[ModelField]) -> ModelNameMap:
-    v1_model_fields = [
-        field for field in fields if isinstance(field, may_v1.ModelField)
-    ]
-    if v1_model_fields:
-        from fastapi._compat import v1
-
-        v1_flat_models = v1.get_flat_models_from_fields(
-            v1_model_fields,  # type: ignore[arg-type]
-            known_models=set(),
-        )
-        all_flat_models = v1_flat_models
-    else:
-        all_flat_models = set()
+    all_flat_models = set()
 
     v2_model_fields = [field for field in fields if isinstance(field, v2.ModelField)]
     v2_flat_models = v2.get_flat_models_from_fields(v2_model_fields, known_models=set())
@@ -187,28 +107,18 @@ def get_definitions(
     model_name_map: ModelNameMap,
     separate_input_output_schemas: bool = True,
 ) -> tuple[
-    dict[
-        tuple[ModelField, Literal["validation", "serialization"]],
-        may_v1.JsonSchemaValue,
-    ],
+    dict[tuple[ModelField, Literal["validation", "serialization"]], dict[str, Any]],
     dict[str, dict[str, Any]],
 ]:
     if sys.version_info < (3, 14):
-        v1_fields = [field for field in fields if isinstance(field, may_v1.ModelField)]
-        v1_field_maps, v1_definitions = may_v1.get_definitions(
-            fields=v1_fields,  # type: ignore[arg-type]
-            model_name_map=model_name_map,
-            separate_input_output_schemas=separate_input_output_schemas,
-        )
-
         v2_fields = [field for field in fields if isinstance(field, v2.ModelField)]
         v2_field_maps, v2_definitions = v2.get_definitions(
             fields=v2_fields,
             model_name_map=model_name_map,
             separate_input_output_schemas=separate_input_output_schemas,
         )
-        all_definitions = {**v1_definitions, **v2_definitions}
-        all_field_maps = {**v1_field_maps, **v2_field_maps}  # type: ignore[misc]
+        all_definitions = {**v2_definitions}
+        all_field_maps = {**v2_field_maps}  # type: ignore[misc]
         return all_field_maps, all_definitions
 
     # Pydantic v1 is not supported since Python 3.14
@@ -228,20 +138,10 @@ def get_schema_from_model_field(
     model_name_map: ModelNameMap,
     field_mapping: dict[
         tuple[ModelField, Literal["validation", "serialization"]],
-        may_v1.JsonSchemaValue,
+        dict[str, Any],
     ],
     separate_input_output_schemas: bool = True,
 ) -> dict[str, Any]:
-    if isinstance(field, may_v1.ModelField):
-        from fastapi._compat import v1
-
-        return v1.get_schema_from_model_field(
-            field=field,
-            model_name_map=model_name_map,
-            field_mapping=field_mapping,
-            separate_input_output_schemas=separate_input_output_schemas,
-        )
-
     return v2.get_schema_from_model_field(
         field=field,  # type: ignore[arg-type]
         model_name_map=model_name_map,
